Добавил черновой вариант создания .pyd файла для нарезчика
