+++ API_cpp_py/src/cpp/cutting_telegrams/tg_cutter/README.md
@@ -1,0 +1,12 @@
+                       Отчет
+За прошедшие две недели я переделал мою функцию разделения на телеграммы.
+
+Основные изменения:
+1. телеграммы достаются рандомно 
+2. учтена проблема большого файла
+3. сделал проверку возможности генерации заданного количества телеграмм
+
+Пока что программа до сих пор не хавает русский язык.
+Генерация позиций для выдергивания телеграмм происходит с помощью библиотеки <random>. 
+Инициализация генератора псевдослучайных чисел gen с помощью Mersenne Twister (MT19937), который является одним из наиболее широко используемых генераторов для равномерного распределения. rd() передается как seed, что делает генератор gen "случайным" при каждом запуске программы.
+
